requirements_dev.txt
bentoml
#bentoml==1.0.10
joblib
pip-chill
torchvision
tqdm
wincertstore
dvc
mlflow
ipykernel
pandas
numpy
seaborn
pytest
tox
black
flake8
mypy
#doto3
streamlit
#-e .
joblib
tqdm
pip-chill
wincertstore
mypy-boto3-s3
torch
torchvision
pillow
Flask
gradio